requirements.txt
numpy
pandas
python_dotenv # To read .env file
mysql-connector-python
pymysql
scikit-learn
seaborn
catboost==1.1.1
xgboost
flask
# -e .